requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-genai>=1.0.0
moviepy>=2.0.0
librosa>=0.10.0
mediapipe>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0
sse-starlette>=2.0.0
aiofiles>=23.0.0
opencv-python-headless>=4.8.0
soundfile>=0.12.0
